LPS-73193 petra-json-ws-client - Reorganize - move classes that shouldn't be exported to internal package
