+++ modules/apps/foundation/petra/petra-json-web-service-client/src/main/java/com/liferay/petra/json/web/service/client/internal/jcifs/JCIFSNTLMSchemeFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.petra.json.web.service.client.jcifs;
+package com.liferay.petra.json.web.service.client.internal.jcifs;
 
 import org.apache.http.auth.AuthScheme;
 import org.apache.http.auth.AuthSchemeProvider;
